backend/requirements.txt
# Core FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.2
requests>=2.31.0

# AI and LLM Integration
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Vector Database
faiss-cpu>=1.7.4

# Voice and Audio Processing
openai-whisper>=20231117
pydub>=0.25.1
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0

# Azure Integration
azure-cognitiveservices-speech>=1.34.0
azure-storage-blob>=12.19.0

# Data Processing and Utilities
pydantic>=2.5.0
typing-extensions>=4.8.0
pandas>=2.0.0
sqlalchemy>=2.0.0

# Development and Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0